♻️ [MOOV-1571]: Moved dialogs and toasts to UI table component
