+++ src/components/functional/PaymentRequestTable/PaymentRequestTable.tsx
@@ -1,12 +1,11 @@
 import { useState } from 'react';
 import { usePaymentRequests } from '../../../api/hooks/usePaymentRequests';
-import { useDeletePaymentRequest } from '../../../api/hooks/useDeletePaymentRequest';
 import UIPaymentRequestTable from '../../ui/PaymentRequestTable/PaymentRequestTable';
 import { RemotePaymentRequestToLocalPaymentRequest } from '../../../utils/parsers';
 import { SortDirection } from '../../ui/ColumnHeader/ColumnHeader';
 import { LocalPaymentRequest } from '../../../api/types/LocalTypes';
-import { Toaster, makeToast } from '../../ui/Toast/Toast';
-import CustomDialog from '../../ui/Dialog/Dialog';
+import { makeToast } from '../../ui/Toast/Toast';
+import { PaymentRequestClient } from '../../../api/clients/PaymentRequestClient';
 
 interface PaymentRequestTableProps {
   token: string; // Example: "eyJhbGciOiJIUz..."
@@ -20,11 +19,8 @@ const PaymentRequestTable = ({ token, apiUrl = 'https://api.nofrixion.com/api/v1
   const [contactSortDirection, setContactSortDirection] = useState<SortDirection>(SortDirection.NONE);
   const [amountSortDirection, setAmountSortDirection] = useState<SortDirection>(SortDirection.NONE);
 
-  /* Deletion confirmation dialog */
-  const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false);
-  const [paymentRequestIdToDelete, setPaymentRequestIdToDelete] = useState('');
-
   const pageSize = 10;
+  const client = new PaymentRequestClient(apiUrl, token);
 
   const { paymentRequests, totalRecords, fetchPaymentRequests } = usePaymentRequests(
     apiUrl,
@@ -37,15 +33,21 @@ const PaymentRequestTable = ({ token, apiUrl = 'https://api.nofrixion.com/api/v1
     pageSize,
   );
 
-  const { deletePaymentRequest } = useDeletePaymentRequest(apiUrl, token);
-
   const localPaymentRequests: LocalPaymentRequest[] = paymentRequests.map((paymentRequest) =>
     RemotePaymentRequestToLocalPaymentRequest(paymentRequest),
   );
 
   const onDeletePaymentRequest = async (paymentRequest: LocalPaymentRequest) => {
-    setPaymentRequestIdToDelete(paymentRequest.id);
-    setIsDeleteDialogOpen(true);
+    const response = await client.delete(paymentRequest.id);
+
+    if (response.error) {
+      makeToast('error', response.error.title);
+      return;
+    }
+
+    makeToast('success', 'Payment request successfully deleted.');
+
+    await fetchPaymentRequests();
   };
 
   const onCopyPaymentRequestLink = async (paymentRequest: LocalPaymentRequest) => {
@@ -55,54 +57,24 @@ const PaymentRequestTable = ({ token, apiUrl = 'https://api.nofrixion.com/api/v1
     makeToast('success', 'Payment request link copied to clipboard.');
   };
 
-  const closeDeleteDialog = () => {
-    setIsDeleteDialogOpen(false);
-  };
-
-  const confirmDeletePaymentRequest = async () => {
-    if (paymentRequestIdToDelete && paymentRequestIdToDelete !== '') {
-      let deleteError = await deletePaymentRequest(paymentRequestIdToDelete);
-
-      if (deleteError) {
-        makeToast('error', deleteError.title);
-        return;
-      }
-
-      makeToast('success', 'Payment request successfully deleted.');
-
-      await fetchPaymentRequests();
-      setPaymentRequestIdToDelete('');
-    }
-
-    setIsDeleteDialogOpen(false);
+  const onDuplicatePaymentRequest = (paymentRequest: LocalPaymentRequest) => {
+    console.log('Duplicate payment request clicked: ', paymentRequest);
   };
 
   return (
-    <>
-      <UIPaymentRequestTable
-        paymentRequests={localPaymentRequests}
-        pageSize={pageSize}
-        totalRecords={totalRecords}
-        onPageChanged={setPage}
-        setStatusSortDirection={setStatusSortDirection}
-        setCreatedSortDirection={setCreatedSortDirection}
-        setContactSortDirection={setContactSortDirection}
-        setAmountSortDirection={setAmountSortDirection}
-        onPaymentRequestDeleteClicked={onDeletePaymentRequest}
-        onPaymentRequestCopyLinkClicked={onCopyPaymentRequestLink}
-      />
-      <Toaster positionY="bottom" positionX="right" duration={5000} />
-      <CustomDialog
-        isOpen={isDeleteDialogOpen}
-        onClose={closeDeleteDialog}
-        title="Confirm payment request deletion."
-        message="Are you sure you want to delete this payment request?"
-        okButtonText="Delete"
-        cancelButtonText="Cancel"
-        okButtonOnClick={confirmDeletePaymentRequest}
-        cancelButtonOnClick={closeDeleteDialog}
-      />
-    </>
+    <UIPaymentRequestTable
+      paymentRequests={localPaymentRequests}
+      pageSize={pageSize}
+      totalRecords={totalRecords}
+      onPageChanged={setPage}
+      setStatusSortDirection={setStatusSortDirection}
+      setCreatedSortDirection={setCreatedSortDirection}
+      setContactSortDirection={setContactSortDirection}
+      setAmountSortDirection={setAmountSortDirection}
+      onPaymentRequestDuplicateClicked={onDuplicatePaymentRequest}
+      onPaymentRequestDeleteClicked={onDeletePaymentRequest}
+      onPaymentRequestCopyLinkClicked={onCopyPaymentRequestLink}
+    />
   );
 };
